rename webpack's config and update its code
